Made some minor changes
Made some minor changes. Updated references in copy to say killing instead of brutality, for brutality by state and time.

I also started messing with the brutality/killings by time by trying to add some zoom/scroll capabilities so that we can shrink the bar chart to take up a smaller container (that way it lines up well with the time series chart to the left of it).