Hooked up our first widget to Bill's Api
Created a new component to show recent activity. Created this component because it 1. Was an easy way for me to test out pulling in API data into the UI and 2. There was actually a request to have some sort of feed that shows recent reports.

At some point before launch, we'll need to prettify the UI a bit, but wanted to check this in so we have a baseline for  pulling API response data into the UI