Refactor last20Victims.js
Refactor includes adding useState hook to handle server response. In addition, removes simpleReact bar and uses css class call to add overflow bar to component (the simpleReact bars in the this module and index.js created a crowding issue on mobile view. For now, use of this css corrects the crowding issue.